% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_rep.R
\name{build_rep}
\alias{build_rep}
\title{construct pseudo-replicates in single-cell data}
\usage{
build_rep(counts, m, q)
}
\arguments{
\item{counts}{: counts matrix from single-cell experiments, genes by cells}

\item{m}{: replicate size, number of cells in each replicate}

\item{q}{: number of replicates}
}
\value{
a list of pseudo-replicates
}
\description{
construct pseudo-replicates in single-cell data
}
\examples{
library(VICE)
data(cmat)
data(cmeta)
psrep <- build_rep(cmat, 100, 3)
}
