% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmat}
\alias{cmat}
\title{Example data for VICE package}
\format{
##
cmat: a data frame with 1000 rows and 100 columns:
\describe{
count matrix of single cell expression (genes by cells)
}
}
\source{
simulated data with splatter
}
\usage{
cmat
}
\description{
Example data for VICE package
}
\keyword{datasets}
