% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmeta}
\alias{cmeta}
\title{Example data for VICE package}
\format{
##
cmeta: a data frame with 1000 rows and 2 columns:
\describe{
  \item{sample}{ID of samples}
  \item{celltype}{name of cell types}
}
}
\source{
simulated data with splatter
}
\usage{
cmeta
}
\description{
Example data for VICE package
}
\keyword{datasets}
