% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvplot.R
\name{cvplot}
\alias{cvplot}
\title{plot CV vs replicate size}
\usage{
cvplot(cvlist, sampleid, cellid)
}
\arguments{
\item{cvlist}{: a list of CV values calculated from get_cv_for_replicates}

\item{sampleid}{: ID of a sample for plot}

\item{cellid}{: ID of a cell type for plot}
}
\value{
a boxplot of CV in sample
}
\description{
plot CV vs replicate size
}
\examples{
library(VICE)
data(cmat)
data(cmeta)
cvlist <- get_cv_for_replicates(cmeta, cmat, 3)
cvplot(cvlist, "s1", "c1")
}
