% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cv.R
\name{get_cv}
\alias{get_cv}
\title{Caculating coefficient of variation}
\usage{
get_cv(x)
}
\arguments{
\item{x}{a vector of gene expression}
}
\value{
CV value
}
\description{
Caculating coefficient of variation
}
\examples{
library(VICE)
data(cmat)
data(cmeta)
get_cv(cmat[1,])#calculate CV for the first gene
}
