% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cv_for_replicates.R
\name{get_cv_for_replicates}
\alias{get_cv_for_replicates}
\title{calculate coefficient of variation (CV) values across pseudo-replicates in single-cell data}
\usage{
get_cv_for_replicates(meta, counts, q)
}
\arguments{
\item{meta}{: metadata of cells, cells by features, must have sample and cell types coulumns}

\item{counts}{: counts matrix from single-cell experiments, genes by cells}

\item{q}{: number of replicates}
}
\value{
: a list of CV across pseudo-replicates expression for each gene
}
\description{
calculate coefficient of variation (CV) values across pseudo-replicates in single-cell data
}
\examples{
library(VICE)
data(cmat)
data(cmeta)
cvlist<-get_cv_for_replicates(cmeta,cmat,3)
}
