% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_perm.R
\name{get_perm}
\alias{get_perm}
\title{permutation test for CV}
\usage{
get_perm(meta, counts, cvlist, q, k)
}
\arguments{
\item{meta}{: metadata of cells}

\item{counts}{: counts matrix from single-cell experiments, genes by cells}

\item{cvlist}{:cvlist calculated from get_cv_for_replicates}

\item{q}{: number of replicates}

\item{k}{: permutation times}
}
\value{
a list of CVs
}
\description{
permutation test for CV
}
\examples{
library(VICE)
data(cmat)
data(cmeta)
cvlist<-get_cv_for_replicates(cmeta,cmat,3)
permcv<-get_perm(cmeta,cmat,cvlist,3,10)
}
