% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample.R
\name{get_sample}
\alias{get_sample}
\title{select sample with the largest number of cell}
\usage{
get_sample(meta)
}
\arguments{
\item{meta}{: metadata of cells}
}
\value{
a dataframe of the samples with largest number of cells for each cell type
}
\description{
select sample with the largest number of cell
}
\examples{
library(VICE)
data(cmat)
data(cmeta)
a <- get_sample(cmeta)
}
