% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tpr.R
\name{get_tpr}
\alias{get_tpr}
\title{Estinate true positive rate of DE genes based on simulated data}
\usage{
get_tpr(ngene, nsample, de.prob, de.facloc, nperm)
}
\arguments{
\item{ngene}{: number of genes}

\item{nsample}{: number of samples in each DE group}

\item{de.prob}{: DE probability}

\item{de.facloc}{: Differential expression factors are produced from a log-normal distribution. Changing these parameters can result in more or less extreme differences between groups.}

\item{nperm}{: iteration times}
}
\value{
TPR of DE analysis
}
\description{
Estinate true positive rate of DE genes based on simulated data
}
\examples{
library(VICE)
test<-get_tpr(ngene=500,nsample=3,de.prob=0.2,de.facloc=0.05,nperm=5)
}
