% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tpr.R
\name{get_tpr}
\alias{get_tpr}
\title{Estinate true positive rate of DE genes based on simulated data}
\usage{
get_tpr(
  ngene = 2000,
  nsample = 3,
  ncell = 100,
  de.prob = 0.3,
  de.facloc = 0.1,
  de.facScale = 0.1,
  nperm = 10
)
}
\arguments{
\item{ngene}{: number of genes}

\item{nsample}{: number of samples in each DE group}

\item{ncell}{: number of cells per sample}

\item{de.prob}{: DE probability}

\item{de.facloc}{: Location (meanlog) parameter for the differential expression factor log-normal distribution. Changing these parameters can result in more or less extreme differences between groups.}

\item{de.facScale}{: Scale (sdlog) parameter for the differential expression factor log-normal distribution. Changing these parameters can result in more or less extreme differences between groups.}

\item{nperm}{: iteration times}
}
\value{
TPR of DE analysis
}
\description{
Estinate true positive rate of DE genes based on simulated data
}
\examples{
library(VICE)
test <- get_tpr(ngene = 2000, ncell = 100, nsample = 3, de.prob = 0.3, de.facScale = 0.1, de.facloc = 0.1, nperm = 10)
}
